+++ src/controllers/ProfileResource.php
@@ -3,10 +3,15 @@
 namespace roaresearch\yii2\roa\controllers;
 
 use roaresearch\yii2\roa\actions\{ProfileUpdate, ProfileView};
-use yii\{filters\VerbFilter, rest\OptionsAction};
+use yii\{base\Model, filters\VerbFilter, rest\OptionsAction};
 
 class ProfileResource extends \yii\rest\Controller
 {
+    /**
+     * @inheritdoc
+     */
+    public $updateScenario = Model::SCENARIO_DEFAULT;
+
     /**
      * @inheritdoc
      */
@@ -43,7 +48,10 @@ class ProfileResource extends \yii\rest\Controller
     {
         return [
             'view' => ['class' => ProfileView::class],
-            'update' => ['class' => ProfileUpdate::class],
+            'update' => [
+                'class' => ProfileUpdate::class,
+                'scenario' => $this->updateScenario,
+            ],
             'options' => ['class' => OptionsAction::class],
         ];
     }