+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Yii2 ROA Library
 2.0.2 June 4, 2021
 ------------------
 
+- [Bug] Fix `ProfileResource::$updateScenario` being ignored
 - [Bug] Fix issue with cached url rules (Faryshta)
+
 
 2.0.1 October 28, 2020
 ----------------------